refactor(sinking): extract module code to sink.py
Move all class/function definitions out of __init__.py into sink.py
and re-export the public API from __init__.py.
